feat: adds a mock python flask based app
the application uses a set of environment variables from an array and renders them
using a jinja template, this is not intended for production use, this is only
meant to be used for our terraform labs to check out setup.

note that flask is running in an asgi wrapper.

environment variables at present have to be added manually via the configuration path